requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15,<4.0
dj-database-url>=2.1,<3.0
psycopg2>=2.9.9,<3.0
gunicorn>=21.2,<22.0
django-cors-headers>=4.0
djangorestframework-simplejwt>=5.3.0
